+++ src/main/java/com/practice/springboot/webservice/board/BoardDao.java
@@ -1,0 +1,5 @@
+package com.practice.springboot.webservice.board;
+
+public class BoardDao {
+
+}
