+++ src/main/java/com/practice/springboot/webservice/board/BoardDaoRepo.java
@@ -1,0 +1,4 @@
+package com.practice.springboot.webservice.board;
+
+public interface BoardDaoRepo {
+}
