+++ src/main/java/com/practice/springboot/webservice/board/BoardService.java
@@ -1,0 +1,7 @@
+package com.practice.springboot.webservice.board;
+
+import org.springframework.stereotype.Service;
+
+@Service
+public class BoardService {
+}
