--- src/main/java/com/practice/springboot/webservice/board/dao/BoardDao.java
@@ -1,4 +1,0 @@
-package com.practice.springboot.webservice.board.dao;
-
-public class BoardDao {
-}
