--- src/main/java/com/practice/springboot/webservice/board/repo/BoardDaoRepo.java
@@ -1,4 +1,0 @@
-package com.practice.springboot.webservice.board.repo;
-
-public interface BoardDaoRepo {
-}
