+++ src/test/java/com/practice/springboot/webservice/board/BoardControllerTest.java
@@ -1,4 +1,4 @@
-package com.practice.springboot.webservice.board.controller;
+package com.practice.springboot.webservice.board;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
